Refactor solidity-wrapper for reuse across Relayer components
