+++ messaging/relayer/internal/msgobserver/eth/utils_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/consensys/gpact/messaging/relayer/internal/messages"
-	"github.com/consensys/gpact/messaging/relayer/internal/msgobserver/eth/soliditywrappers/sfc"
+	"github.com/consensys/gpact/messaging/relayer/internal/soliditywrappers/sfc"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind/backends"
 	"github.com/ethereum/go-ethereum/common"
